test: mock database url for all tests
