+++ tests/conftest.py
@@ -1,0 +1,4 @@
+import os
+
+# Ensure required environment variables are present for tests
+os.environ.setdefault("DATABASE_URL", "postgresql://fake:fake@localhost:5432/fake")
